+++ binarySearch/binarySearch/binarySearch.vcxproj
@@ -140,11 +140,12 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="binarySearch.c" />
+    <ClCompile Include="main.c" />
     <ClCompile Include="qsort.c" />
-    <ClCompile Include="tests.c" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Header.h" />
+    <ClInclude Include="binary.h" />
+    <ClInclude Include="qsort.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
