+++ binarySearch/binarySearch/binarySearch.vcxproj.filters
@@ -21,12 +21,15 @@
     <ClCompile Include="qsort.c">
       <Filter>Исходные файлы</Filter>
     </ClCompile>
-    <ClCompile Include="tests.c">
+    <ClCompile Include="main.c">
       <Filter>Исходные файлы</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Header.h">
+    <ClInclude Include="qsort.h">
+      <Filter>Файлы заголовков</Filter>
+    </ClInclude>
+    <ClInclude Include="binary.h">
       <Filter>Файлы заголовков</Filter>
     </ClInclude>
   </ItemGroup>
